+++ Lib/test/test_repl.py
@@ -14,7 +14,7 @@ from test.support import (
     SuppressCrashReport,
     SHORT_TIMEOUT,
 )
-from test.support.script_helper import kill_python, spawn_python
+from test.support.script_helper import kill_python
 from test.support.import_helper import import_module
 
 try:
@@ -97,35 +97,6 @@ class TestInteractiveInterpreter(unittest.TestCase):
         self.assertIn('After the exception.', output)
         # Exit code 120: Py_FinalizeEx() failed to flush stdout and stderr.
         self.assertIn(p.returncode, (1, 120))
-
-    @cpython_only
-    # Python built with Py_TRACE_REFS fail with a fatal error in
-    # _PyRefchain_Trace() on memory allocation error.
-    @unittest.skipIf(support.Py_TRACE_REFS, 'cannot test Py_TRACE_REFS build')
-    def test_exec_set_nomemory_hang(self):
-        import_module("_testcapi")
-        # gh-134163: Test case that triggers no memory hang condition
-        # The frame_lasti need to upper 257,
-        # because when calling PyLong_FromLong, malloc is not invoked,
-        # so no MemError is triggered
-        # we need to warm up the memory to reproduce the issue
-        warmup_code = "a = list(range(0, 1))\n" * 20
-        user_input = warmup_code + dedent("""
-            try:
-                import _testcapi
-                _testcapi.set_nomemory(0)
-                b = list(range(1000, 2000))
-            except Exception as e:
-                import traceback
-                traceback.print_exc()
-            """)
-        with SuppressCrashReport():
-            p = spawn_python('-c', user_input)
-        output = kill_python(p)
-
-        self.assertIn(p.returncode, (0, 1, 120))
-        self.assertGreater(len(output), 0)  # At minimum, should not hang
-        self.assertIn(b"MemoryError", output)
 
     @cpython_only
     def test_multiline_string_parsing(self):
